git-svn-id: https://osirix.svn.sourceforge.net/svnroot/osirix@3655 d300eccd-f00d-0410-ae38-fa88e4616176

+++ plugins/CMIV_CTA_TOOLS/CMIVChopperController.h
@@ -60,7 +60,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 	NSRect   roiRect;
 	NSRect   coronalROIRect;
 	NSRect   sagittalROIRect;
-	int      imageWidth,imageHeight,imageAmount;
+	long      imageWidth,imageHeight,imageAmount;
 	int      iImageFrom,iImageTo; 
 	float    ratioXtoThick,ratioYtoThick,ratioXtoY;
 	float    *outputVolumeData;
@@ -72,8 +72,8 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 	
 	ViewerController     *originalViewController;
 	
-	float				vtkOriginalX,vtkOriginalY,vtkOriginalZ;
-	float				sliceThickness;
+	double				vtkOriginalX,vtkOriginalY,vtkOriginalZ;
+	double				sliceThickness;
 	
 	vtkImageReslice		*rotate;
 	vtkImageImport		*reader;
